remove detect browser
