+++ website/package.json
@@ -22,7 +22,6 @@
     "classnames": "^2.3.2",
     "color": "^4.2.3",
     "cookie": "^0.5.0",
-    "detect-browser": "^5.3.0",
     "idb": "^8.0.0",
     "locale-codes": "^1.3.1",
     "lodash.get": "^4.4.2",
